Fragment RingbufferService and support predicates and projections
Adapt RingbufferService to hold ringbuffers for other services (e.g.
map). This means that it should implement
FragmentedMigrationAwareService and return partition namespace data
for other services when requested.

Also, the read result sets were returning only how many items were
read as well as the matching items. In case the ringbuffer is read
with a filter, not all items match and the result set may not contain
items with a contiguous set of sequences. The itemCount is the
difference between the sequence ID of the last and first returned
item and is a hack not to send all sequence IDs.

We will send the sequence IDs instead. This allows the user to store
the sequence ID and resume processing even if the processing crashes
mid-point in the result set.

The ringbuffer already has supports for IFilter but we also add support
for predicates and projections.