Add projection and filtering support for map iterator (#66)
The added projection and filtering support reuses the query engine but
allows running the query on chunk within a single partition. This
allows iteration with predicates and projections. Currently iterating
the indexes is not supported, the predicate will be run on each entry.
Also, the projection does not support the Offloadable which means that
it is invoked on the partition thread.

Changed the type of the entry which is provided to the predicate from
CachedQueryEntry to LazyMapEntry to allow the identity projection -
previously it failed when being serialized since CachedQueryEntry is
not serializable.

Also added some more generics and javadocs.