Add next sequence to read from in ResultSet response (#124)
Sometimes the reader of the ringbuffer is loss tolerant and we may
allow the reader to be slow and skip some items. To indicate this, we
send an additional sequence to indicate where the reader can continue
reading from number of lost items so that the reader may check if
items have been lost (overwritten) and act on that information.