Put newly added fix sized params to begin frame
We don't encode an initial frame if the custom type does not
have any fix sized parameters. However, if the custom type
is extended with a new fix sized parameter, new codecs
will try to encode/decode an initial frame which will
broke the communication between old clients/members
since they don't expect an initial frame. Therefore,
if the custom type is extended with a fix sized parameter,
we encode the data into the begin frame