[BACKPORT] Add codec versioning support (#310)
* add ability to add new parameters to codecs

* update implementation

* Removed the hasNext check from the custom codec tempalate
since custom codecs always have an end frame at the end.
We should only check whether or not the next frame
is an end frame or not.

* Put newly added fix sized params to begin frame

We don't encode an initial frame if the custom type does not
have any fix sized parameters. However, if the custom type
is extended with a new fix sized parameter, new codecs
will try to encode/decode an initial frame which will
broke the communication between old clients/members
since they don't expect an initial frame. Therefore,
if the custom type is extended with a fix sized parameter,
we encode the data into the begin frame

* relax schma validation for the since values and check ordering of parameters