Update the generator to use the new since versions
The new since versions are now tied up to the Hazelcast version.

This commit updates the documentation and schemas with the new
structure of the since versions.

Also, this removes the checks for the subsequent protocol versions,
as it is now OK to skip some since versions, if there are no
changes in some Hazelcast versions.